Safari-609.1.11.1/LayoutTests/media/video-orientation-expected.txt


Test size before movie is open:
EXPECTED (video.videoWidth == '0') OK
EXPECTED (video.videoHeight == '0') OK

Loading "content/no-rotation.mp4". Movie video with no rotation tag set, size 352x288.
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EXPECTED (video.videoWidth == '352') OK
EXPECTED (video.videoHeight == '288') OK

Loading "content/rotation-90.mp4". Movie video with rotation-90 tag set, size 352x288.
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EXPECTED (video.videoWidth == '288') OK
EXPECTED (video.videoHeight == '352') OK

Loading "content/rotation-180.mp4". Movie video with rotation-180 tag set, size 352x288.
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EXPECTED (video.videoWidth == '352') OK
EXPECTED (video.videoHeight == '288') OK

Loading "content/rotation-270.mp4". Movie video with rotation-270 tag set, size 352x288.
EVENT(loadstart)
EVENT(durationchange)
EVENT(loadedmetadata)
EXPECTED (video.videoWidth == '288') OK
EXPECTED (video.videoHeight == '352') OK

END OF TEST
